# services/gateway/docs/swagger.yaml
definitions:
  github_com_ummuys_reportify_services_gateway_internal_webdto.BaseResponse:
    properties:
      msg:
        type: string
    type: object
  github_com_ummuys_reportify_services_gateway_internal_webdto.CreateReportRequest:
    properties:
      comment:
        type: string
      csv_separator:
        type: string
      format:
        type: string
      name:
        type: string
      query_sql:
        type: string
    type: object
  github_com_ummuys_reportify_services_gateway_internal_webdto.CreateReportResponse:
    properties:
      report_id:
        type: string
      status:
        type: string
    type: object
...
  github_com_ummuys_reportify_services_gateway_internal_webdto.DeleteReportResponse:
    properties:
      report_id:
        type: string
    type: object
...
  github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse:
    properties:
      err:
        type: string
    type: object
  github_com_ummuys_reportify_services_gateway_internal_webdto.ListReportsResponse:
    properties:
      reports:
        items:
          $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ReportMetadata'
        type: array
    type: object
  github_com_ummuys_reportify_services_gateway_internal_webdto.ListSchemasResponse:
    properties:
      schemas:
        items:
          $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.Schema'
        type: array
    type: object
  github_com_ummuys_reportify_services_gateway_internal_webdto.ListTablesResponse:
    properties:
      tables:
        items:
          $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.Table'
        type: array
    type: object
...
  github_com_ummuys_reportify_services_gateway_internal_webdto.ReportInfoResponse:
    properties:
      report:
        $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ReportMetadata'
    type: object
  github_com_ummuys_reportify_services_gateway_internal_webdto.ReportMetadata:
    properties:
      author_id:
        type: string
      comm:
        type: string
      created_at:
        type: string
      csv_sep:
        type: string
      err_msg:
        type: string
      file_path:
        type: string
      format:
        type: string
      name:
        type: string
      query:
        type: string
      report_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  github_com_ummuys_reportify_services_gateway_internal_webdto.ReportStatusResponse:
    properties:
      report_id:
        type: string
      status:
        type: string
    type: object
  github_com_ummuys_reportify_services_gateway_internal_webdto.Schema:
    properties:
      comment:
        type: string
      name:
        type: string
    type: object
  github_com_ummuys_reportify_services_gateway_internal_webdto.Table:
    properties:
      comment:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /reports:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
      security:
      - BearerAuth: []
      summary: Delete all user reports
      tags:
      - reports
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ListReportsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
      security:
      - BearerAuth: []
      summary: List user reports
      tags:
      - reports
    post:
      consumes:
      - application/json
      parameters:
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.CreateReportRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.CreateReportResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
      security:
      - BearerAuth: []
      summary: Create report
      tags:
      - reports
  /reports/{report_id}:
    delete:
      parameters:
      - description: Report ID
        in: path
        name: report_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.DeleteReportResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
      security:
      - BearerAuth: []
      summary: Delete one report
      tags:
      - reports
    get:
      parameters:
      - description: Report ID
        in: path
        name: report_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ReportInfoResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
      security:
      - BearerAuth: []
      summary: Get report info
      tags:
      - reports
  /reports/{report_id}/status:
    get:
      parameters:
      - description: Report ID
        in: path
        name: report_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ReportStatusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
      security:
      - BearerAuth: []
      summary: Get report status
      tags:
      - reports
  /schemas:
    get:
      description: Returns list of available schemas
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ListSchemasResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
      security:
      - BearerAuth: []
      summary: List schemas
      tags:
      - datasource
  /tables:
    get:
      description: Returns list of tables for given schema
      parameters:
      - description: Schema name
        in: query
        name: schema
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ListTablesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
      security:
      - BearerAuth: []
      summary: List tables in schema
      tags:
      - datasource
...
swagger: "2.0"